Update countdown formatting to use f64 instead of i64
- Changed the parameter type of `format_countdown` from `i64` to `f64` for improved precision in countdown calculations.
- Updated the calculation of `secs` in `ProjectRow` to reflect the new type, ensuring consistent handling of time values.

+++ frontend/src/components/dashboard.rs
@@ -12,7 +12,7 @@ use wasm_bindgen::prelude::*;
 #[wasm_bindgen]
 extern "C" {
     #[wasm_bindgen(js_name = formatCountdown)]
-    fn format_countdown(secs: i64) -> String;
+    fn format_countdown(secs: f64) -> String;
 }
 
 // --- Logic Layer: Dashboard Store ---
@@ -443,7 +443,7 @@ fn ProjectRow(project: ProjectConfig) -> impl IntoView {
             MonitorState::Paused => "--".to_string(),
             MonitorState::Running { next_check_at } => {
                 let now = Date::now_timestamp();
-                let secs = (*next_check_at - now).as_secs() as i64;
+                let secs = (*next_check_at - now).as_secs() as f64;
                 format_countdown(secs)
             }
         }
